Updated structure Added all functionality Needs test and fixes
Changes:
	moved and changed:      ua/lpnu/students/labs/laba2/ChristmassDecoration/Manager.java -> ua/lpnu/students/labs/laba2/ChristmassDecoration/Manager/Manager.java
	moved and changed:      ua/lpnu/students/labs/laba2/ChristmassDecoration/ElectricDecoration.java -> ua/lpnu/students/labs/laba2/ChristmassDecoration/Model/ElectricDecoration.java
	moved and changed:      ua/lpnu/students/labs/laba2/ChristmassDecoration/LongDecoration.java -> ua/lpnu/students/labs/laba2/ChristmassDecoration/Model/LongDecoration.java
	moved and changed:      ua/lpnu/students/labs/laba2/ChristmassDecoration/OrganicDecoration.java -> ua/lpnu/students/labs/laba2/ChristmassDecoration/Model/OrganicDecoration.java
	moved and changed:      ua/lpnu/students/labs/laba2/ChristmassDecoration/PieceDecoration.java -> ua/lpnu/students/labs/laba2/ChristmassDecoration/Model/PieceDecoration.java
	moved and changed:      ua/lpnu/students/labs/laba2/ChristmassDecoration/Template.java -> ua/lpnu/students/labs/laba2/ChristmassDecoration/Model/shared/Template.java
	moved and changed:      ua/lpnu/students/labs/laba2/ChristmassDecoration/Type.java -> ua/lpnu/students/labs/laba2/ChristmassDecoration/Model/shared/Type.java
	moved and changed:      ua/lpnu/students/labs/laba2/TextMenu.java -> ua/lpnu/students/labs/laba2/ChristmassDecoration/TextMenu/TextMenu.java
	moved and changed:      ua/lpnu/students/labs/laba2/Main.java -> ua/lpnu/students/labs/laba2/ChristmassDecoration/tests/Main.java
