--- ua/lpnu/students/labs/laba2/ChristmassDecoration/Type.java
@@ -1,5 +1,0 @@
-package ua.lpnu.students.labs.laba2.ChristmassDecoration;
-
-public enum Type {
-    OUTSIDE_DECORATION, FOR_CHRISTMASS, UNIVERSAL, FOR_WINDOWS,
-}
